Rename jury_description file to jury-description
