+++ src/components/jury-description.jsx
