+++ src/components/jury.jsx
@@ -4,8 +4,8 @@ import React from "react";
 import { Grid, Col, Row } from "react-bootstrap";
 
 import Header from "./header";
-import JuryDescription from "./jury_description.jsx";
-import Authentication from "../mixins/authentication.jsx";
+import JuryDescription from "./jury-description";
+import Authentication from "../mixins/authentication";
 
 import "./jury.less";
